+++ mini-projects/day05-indexing.py
@@ -1,0 +1,31 @@
+#Ask user for a sentence
+#↓
+#Print Original
+#↓
+#Print UPPERCASE
+#↓
+#Print lowercase
+#↓
+#Print Title Case
+#↓
+#Print Character Count
+#↓
+#Print First Character
+#↓
+#Print Last Character
+#↓
+#Replace one word with another
+#↓
+#Check if a word exists
+
+sentence = "I love myself"
+print(sentence)
+print(sentence.upper())
+print(sentence.lower())
+print(sentence.title())
+print(len(sentence))
+print(sentence[0])
+print(sentence[-1])
+print(sentence.replace("myself","everyone"))
+print(sentence.find("love"))
+print("love" in sentence )