+++ while-loop/while-loop-tes01.py
@@ -1,4 +1,0 @@
-count = 1
-while count<=1:
-    print(count)
-    count = count + 1