+++ while-loop/while-loop-test01.py
@@ -1,0 +1,7 @@
+number = 7
+guess = 0
+
+while guess != number:
+    guess = int(input("Enter your guess: "))
+
+print("🎉 Correct!")